requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorgram.py==1.2.0
colorthief==0.2.1
convcolors==1.0.0
cycler==0.10.0
Django==3.1.7
extcolors==1.0.0
idna==2.10
image==1.5.33
ImageHash==4.2.0
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
scipy==1.6.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
